Feature: move Restqapi to the monorepo. And rename restqapi to rest-api
