+++ docs/content/test-creation/given/authorization.md
@@ -4,7 +4,7 @@ title: Authorization
 sidebar_label: Authorization
 parent: given
 category: Authorization
-content_from: restqapi/src/restqapi/steps/1-given/index.js
+content_from: rest-api/src/rest-api/steps/1-given/index.js
 order: 4
 ---
