+++ docs/content/test-creation/given/host.md
@@ -4,7 +4,7 @@ title: Host
 sidebar_label: Host
 parent: given
 category: Host
-content_from: restqapi/src/restqapi/steps/1-given/index.js
+content_from: rest-api/src/rest-api/steps/1-given/index.js
 order: 1
 ---
