+++ docs/content/test-creation/given/method.md
@@ -4,7 +4,7 @@ title: Request method
 sidebar_label: Request method
 parent: given
 category: Method
-content_from: restqapi/src/restqapi/steps/1-given/index.js
+content_from: rest-api/src/rest-api/steps/1-given/index.js
 order: 3
 ---
