+++ docs/content/test-creation/given/path.md
@@ -4,7 +4,7 @@ title: Request URL path
 sidebar_label: URL path
 parent: given
 category: Path
-content_from: restqapi/src/restqapi/steps/1-given/index.js
+content_from: rest-api/src/rest-api/steps/1-given/index.js
 order: 2
 ---
