+++ docs/content/test-creation/given/query.md
@@ -4,7 +4,7 @@ title: Request Query Parameter
 sidebar_label: Query parameter
 parent: given
 category: Query string
-content_from: restqapi/src/restqapi/steps/1-given/index.js
+content_from: rest-api/src/rest-api/steps/1-given/index.js
 order: 5
 ---
