+++ docs/content/test-creation/given/request-body-form.md
@@ -4,7 +4,7 @@ title: Form Request Body
 sidebar_label: Form Request Body
 parent: given
 category: Form Request body
-content_from: restqapi/src/restqapi/steps/1-given/index.js
+content_from: rest-api/src/rest-api/steps/1-given/index.js
 order: 7
 ---
