+++ docs/content/test-creation/given/request-body-json.md
@@ -4,7 +4,7 @@ title: JSON Request Body
 sidebar_label: JSON Request Body
 parent: given
 category: JSON Request body
-content_from: restqapi/src/restqapi/steps/1-given/index.js
+content_from: rest-api/src/rest-api/steps/1-given/index.js
 order: 6
 ---
