+++ docs/content/test-creation/then/contract-validation.md
@@ -4,7 +4,7 @@ title: Response Contract Validation
 sidebar_label: Contract validation
 parent: then
 category: Validation
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 6
 ---
