+++ docs/content/test-creation/then/cookies.md
@@ -4,7 +4,7 @@ title: Response Cookie
 sidebar_label: Cookie
 parent: then
 category: Cookie
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 7
 ---
