+++ docs/content/test-creation/then/dataset.md
@@ -4,7 +4,7 @@ title: Response Dataset
 sidebar_label: Dataset
 parent: then
 category: Dataset
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 7
 ---
