+++ docs/content/test-creation/then/date.md
@@ -4,7 +4,7 @@ title: Response Date
 sidebar_label: Date
 parent: then
 category: Date
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 6
 ---
