+++ docs/content/test-creation/then/headers.md
@@ -4,7 +4,7 @@ title: Response Headers
 sidebar_label: Headers
 parent: then
 category: Headers
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 3
 ---
