+++ docs/content/test-creation/then/json-body.md
@@ -4,7 +4,7 @@ title: Response JSON
 sidebar_label: JSON Body
 parent: then
 category: JSON Response body
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 4
 ---
