+++ docs/content/test-creation/then/sort-numeric.md
@@ -4,7 +4,7 @@ title: Response Sort Numeric
 sidebar_label: Sort numeric
 parent: then
 category: Sort Numeric
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 5
 ---
