+++ docs/content/test-creation/then/status.md
@@ -4,7 +4,7 @@ title: Response Status
 sidebar_label: HTTP code
 parent: then
 category: Status code
-content_from: restqapi/src/restqapi/steps/3-then/index.js
+content_from: rest-api/src/rest-api/steps/3-then/index.js
 order: 1
 ---
