+++ docs/content/test-creation/when/send.md
@@ -4,7 +4,7 @@ title: Sending your request
 sidebar_label: Send
 parent: when
 category: Send
-content_from: restqapi/src/restqapi/steps/2-when/index.js
+content_from: rest-api/src/rest-api/steps/2-when/index.js
 order: 1
 ---
