+++ package.json
@@ -15,12 +15,14 @@
     "docs",
     "ui",
     "cli",
-    "example"
+    "example",
+    "plugins/rest-api"
   ],
   "dependencies": {
     "@restqa/docs": "file:docs",
     "@restqa/ui": "file:ui",
-    "@restqa/restqa": "file:restqa"
+    "@restqa/restqa": "file:restqa",
+    "@restqa/rest-a": "file:plugins/rest-api"
   },
   "engines": {
     "npm": ">=8.0.0",
