+++ plugins/rest-api/example/setup.js
