+++ plugins/rest-api/package.json
@@ -1,10 +1,9 @@
 {
-  "name": "@restqa/restqapi",
+  "name": "@restqa/rest-api",
   "version": "0.1.5",
   "private": false,
   "description": "REST API Functional Test engine based on Gerkin - RestQA",
   "scripts": {
-    "doc": "jsdoc2md  --partial docs/support/scope.hbs --partial docs/support/header.hbs --files src/restqapi/steps/**/index.js > docs/steps-catalog.md",
     "preexample": "npm run build",
     "example": "cucumber-js --require ./example/setup.js --publish-quiet --fail-fast ./example/features",
     "format": "prettier --check src",
@@ -40,7 +39,6 @@
     "jest": "^26.6.3",
     "jest-runner-eslint": "^0.10.0",
     "jest-runner-prettier": "^0.3.7",
-    "jsdoc-to-markdown": "^5.0.3",
     "nodemon": "^2.0.3",
     "prettier": "^2.4.1"
   },
