+++ plugins/rest-api/src/rest-api/lib/generator.js
