+++ plugins/rest-api/src/rest-api/lib/index.test.js
